Copies code and adapts to updated dependencies
It splits the code between `src/main/` and `src/test/`.  It updates the
examples to the latest JMock, MockEasy, and Mockito.  It does **not**
update them to JUnit 4.

JMock and EasyMock have not had a whole lot of development in the past
ten years, so I'm confident the examples still hold true.  For the most
part.  Mockito has had a lot of development and has a lot of new
features.  I suspect that many areas where it used to be "n/a" are now
fully covered by that framework.  I'll have to look  it up.